+++ src/app/events/[slug]/components/SponsorsSection.tsx
@@ -1,38 +1,19 @@
 'use client'
 
-import Image from 'next/image'
-
-import { clsx } from 'clsx'
-
 import type { Event } from '@/types/eventType'
 
 import { buildImageSizeProp } from '@/utils/buildImageSizeProp'
-import { capitalize } from '@/utils/capitalize'
 
 import { PageSection } from '@/components/PageSection'
 
-type Sponsor = NonNullable<NonNullable<Event['sponsors']>['first-tier']>[number]
+import { SponsorGrid } from './SponsorGrid'
+
+export type Sponsor = NonNullable<
+  NonNullable<Event['sponsors']>['first-tier']
+>[number]
 
 type SponsorsSectionProps = {
   sponsors: Event['sponsors']
-}
-
-type SponsorGridProps = {
-  sponsors: Array<Sponsor>
-  tier: string
-  columnsStyle: string
-  imageConfig: ImageConfig
-}
-
-type SponsorLogoProps = {
-  sponsor: Sponsor
-  tier: string
-  imageConfig: ImageConfig
-}
-
-type ImageConfig = {
-  containerStyles: string
-  imageSizes: string
 }
 
 const sponsorTierConfigs = {
@@ -88,50 +69,5 @@ export function SponsorsSection({ sponsors }: SponsorsSectionProps) {
         })}
       </div>
     </PageSection>
-  )
-}
-
-function SponsorGrid({
-  sponsors,
-  tier,
-  columnsStyle,
-  imageConfig,
-}: SponsorGridProps) {
-  return (
-    <div className={clsx('grid gap-8', columnsStyle)}>
-      {sponsors.map((sponsor, index) => (
-        <SponsorLogo
-          key={index}
-          sponsor={sponsor}
-          tier={tier}
-          imageConfig={imageConfig}
-        />
-      ))}
-    </div>
-  )
-}
-function SponsorLogo({ sponsor, tier, imageConfig }: SponsorLogoProps) {
-  const formattedTier = capitalize(tier.replace('-', ' '))
-
-  return (
-    <div className={clsx('relative w-full', imageConfig.containerStyles)}>
-      <Image
-        key={sponsor.name}
-        src={sponsor.image.src}
-        alt={`${sponsor.name} Logo - ${formattedTier} sponsor`}
-        fill={true}
-        sizes={imageConfig.imageSizes}
-        className="object-contain object-left-bottom"
-      />
-      {sponsor.website && (
-        <a
-          href={sponsor.website}
-          target="_blank"
-          rel="noopener noreferrer"
-          aria-label={`Visit ${sponsor.name}'s website`}
-          className="absolute inset-0 rounded-lg focus:brand-outline"
-        />
-      )}
-    </div>
   )
 }
